+++ amplify/functions/attendees/resource.ts
@@ -1,7 +1,0 @@
-import { defineFunction } from '@aws-amplify/backend';
-
-export const attendeesFn = defineFunction({
-  name: 'attendees',
-  entry: '../../../backend-aws/functions/attendees/index.js',
-  runtime: 20,
-});